% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_images.R
\name{check_images}
\alias{check_images}
\title{Check images}
\usage{
check_images(lesson_file, images_folder, ignore_extensions = FALSE)
}
\arguments{
\item{lesson_file}{the .md file with the lesson you want to check}

\item{images_folder}{the folder where the images of the
lesson are stored}

\item{ignore_extensions}{whether you want to ignore checking the
extensions of the image files or not. Default value is FALSE,
and that is a good idea, unless the lesson you are editing
has a very good reason to use other formats.}
}
\description{
\code{check_images()} checks if images in a lesson file comply
Programming Historian editorial guidelines. It will check:
1. if images names in the lesson file match the names of
files in the images folder. If not, it will return
names of images producing the mismatch.
2. if images has the lesson slug as name.
3. if images has the expected extensions.
}
\examples{
\dontrun{
check_images(lesson_file = "awesome-new-lesson.md",
    images_folder = "awesome-new-lesson")}

}
