% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_links.R
\name{check_links}
\alias{check_links}
\title{Check if lesson links are active}
\usage{
check_links(lesson_file)
}
\arguments{
\item{lesson_file}{path to the .md file you want to check (or just
the name if it is on your working directory)}
}
\value{
a data frame with status of each checked link
}
\description{
Check if lesson links are active
}
\examples{
\dontrun{check_links("awesome-new-lesson.md")}
}
