% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open_guidelines.R
\name{open_guidelines}
\alias{open_guidelines}
\alias{open_editor_guidelines}
\alias{open_reviewer_guidelines}
\alias{open_translator_guidelines}
\alias{open_author_guidelines}
\title{Open PH guidelines in your default browser}
\usage{
open_editor_guidelines(language)

open_reviewer_guidelines(language)

open_translator_guidelines(language)

open_author_guidelines(language)
}
\arguments{
\item{language}{the language of the guidelines page you want to open}
}
\description{
\verb{open_*_guidelines()} functions allow you to open any editorial
guidelines in the language of your preference
}
\examples{
\dontrun{
open_author_guidelines("en")
open_editor_guidelines("es")
open_reviewer_guidelines("fr")
open_translator_guidelines("pt")
}
}
