% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_images.R
\name{rename_images}
\alias{rename_images}
\title{Rename images to match PH guidelines}
\usage{
rename_images(lesson_file, images_folder, backup_files = TRUE)
}
\arguments{
\item{lesson_file}{the .md file with the lesson you want to edit}

\item{images_folder}{the folder where the images of the
lesson are stored}

\item{backup_files}{whether you want to keep a copy of the
original files or not. Default value is TRUE.}
}
\description{
Rename images in the .md and files in the images folder
for them to comply Programming Historian editorial
guidelines.
}
\examples{
\dontrun{
rename_images(lesson_file = "awesome-new-lesson.md",
    images_folder = "awesome-new-lesson")}
}
